requirements.txt
# Core dependencies
boto3>=1.34.0
botocore>=1.34.0

# GitHub API
PyGithub>=2.1.0

# Security scanning
checkov>=3.0.0
tfsec>=1.0.0

# Terraform/CDK
terraform-compliance>=1.3.0

# Utilities
requests>=2.31.0
pyyaml>=6.0
jsonschema>=4.19.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
moto>=4.2.0

# Code quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
isort>=5.12.0

# Security
safety>=2.3.0
bandit>=1.7.0
